+++ src/plugins/contentRenderer/bodyHeader/cssDivs/cssDivsContentRenderer.component.ts
@@ -5,10 +5,9 @@ import {isArray} from '@jscrpt/common';
 
 import {BodyHeaderContentRendererAbstractComponent} from '../bodyHeaderContentRendererAbstract.component';
 import {CssDivsContentRendererOptions} from './cssDivsContentRenderer.interface';
-import {GridMetadata, TableGridColumn} from '../../../../interfaces';
+import {GridMetadata, TableGridColumn, TableGridMetadata} from '../../../../interfaces';
 import {CONTENT_RENDERER_OPTIONS, GRID_PLUGIN_INSTANCES} from '../../../../misc/tokens';
 import {GridPluginInstances} from '../../../../misc/types';
-import {TableGridMetadata} from '../../../../components/tableGridMetadataGatherer/tableGridMetadataGatherer.interface';
 import {CssDivsBodyContentRendererSAComponent} from './body/cssDivsBodyContentRenderer.component';
 import {CssDivsHeaderContentRendererSAComponent} from './header/cssDivsHeaderContentRenderer.component';
 import {ResolveForwardRefSAPipe} from '../../../../pipes';