+++ src/plugins/contentRenderer/bodyHeader/table/body/basic/tableBodyContentRenderer.component.ts
@@ -3,8 +3,7 @@ import {CommonModule} from '@angular/common';
 
 import {TableBodyContentRendererOptions} from '../../tableContentRenderer.interface';
 import {BodyContentRendererAbstractComponent} from '../../../bodyContentRendererAbstract.component';
-import {TableGridMetadata} from '../../../../../../components/tableGridMetadataGatherer/tableGridMetadataGatherer.interface';
-import {TableGridColumn} from '../../../../../../interfaces';
+import {TableGridColumn, TableGridMetadata} from '../../../../../../interfaces';
 import {BODY_CONTENT_RENDERER_OPTIONS, GRID_PLUGIN_INSTANCES} from '../../../../../../misc/tokens';
 import {GridPluginInstances} from '../../../../../../misc/types';
 import {DataCellContextSAPipe, ReadValueSAPipe} from '../../../../../../pipes';